feat: Add Sarah’s additions.
This comment adds Sarah’s files to the source tree.

The files are:
- invenio_cli/commands/translations.py
- invenio_cli/cli/translations.py

Note: The repository is not organized in any way. The whole repository needs to be integrated to the bigger InvenioRDM ecosystem. Do not use this code as-is.

Note: The code adds a new dependency (polib), and it’s important to make sure that it doesn’t clash with anything in the bigger project.